[WIP] Watch view implementation to track running 'odo watch' commands
Watch commands can be executed for different components at the same time
and that is done in terminal veiw. Over time terminal is getting
polluted with different commands and it is getting not easy to nawigate
and search for watch sessions to see the log or stop them.

This fix shows all components under watch in a view so devs can:
1. See the logs
2. Stop active watch commands

Related to #1425.

+++ src/view/log/LogViewLoader.ts
@@ -9,6 +9,7 @@ import { ExtenisonID } from '../../util/constants';
 import { OpenShiftObject } from '../../odo';
 import * as odo from '../../odo';
 import treeKill = require('tree-kill');
+import { ChildProcess } from 'child_process';
 
 export default class LogViewLoader {
 
@@ -16,7 +17,7 @@ export default class LogViewLoader {
         return vscode.extensions.getExtension(ExtenisonID).extensionPath
     }
 
-    static async loadView(title: string, cmdFunction: (prj, app, comp) => string, target: OpenShiftObject): Promise<vscode.WebviewPanel> {
+    static async loadView(title: string, cmdFunction: (prj, app, comp) => string, target: OpenShiftObject, existingProcess?: ChildProcess): Promise<vscode.WebviewPanel> {
         const localResourceRoot = vscode.Uri.file(path.join(LogViewLoader.extensionPath, 'out', 'logViewer'));
 
         const panel = vscode.window.createWebviewPanel('logView', title, vscode.ViewColumn.One, {
@@ -31,7 +32,7 @@ export default class LogViewLoader {
         // TODO: When webview is going to be ready?
         panel.webview.html = LogViewLoader.getWebviewContent(LogViewLoader.extensionPath, cmd);
 
-        const process = await odo.getInstance().spawn(cmd, target.contextPath.fsPath);
+        const process = existingProcess? existingProcess : await odo.getInstance().spawn(cmd, target.contextPath.fsPath);
         process.stdout.on('data', (data) => {
             panel.webview.postMessage({action: 'add', data: `${data}`.trim().split('\n')});
         }).on('close', ()=>{
@@ -43,9 +44,11 @@ export default class LogViewLoader {
                 recieveDisposable.dispose();
             }
         })
-        const disposable = panel.onDidDispose(()=> {
-            treeKill(process.pid);
-            disposable.dispose();
+        panel.onDidDispose(()=> {
+            process.stdout.removeAllListeners();
+            if(!existingProcess) {
+                treeKill(process.pid);
+            }
         });
         return panel;
     }
